Separate binder (element to bind onscroll) and target (element to look at bottom pixel)
Use case is when the overflow:auto element (binder) is already taller
(due to unrelated content) than the element (target) that holds the
collection items. (Imagine a 2 column layout).
You want it to load in more items as you scroll to the
bottom of the 'target' even if you're nowhere near the bottom of
'binder'.